Replace Orignal TransactionContext with ZmqTransactionContext
This includes updating the context methods to use
references and properly handle get_state with multiple
addresses. Also adds get_state_entry and get_state_entries
to the ZmqTransactionContext.

This commit also moves the ZmqTransactionContext into its own
private mod, since SDK users should not be creating one of these
directly
